requirements.txt
alembic==1.0.11
Click==7.0
dominate==2.4.0
Flask==1.1.1
Flask-Bootstrap==3.3.7.1
flask-marshmallow==0.10.1
Flask-Migrate==2.5.2
Flask-SQLAlchemy==2.4.0
Flask-WTF==0.14.2
itsdangerous==1.1.0
Jinja2==2.10.1
Mako==1.1.0
MarkupSafe==1.1.1
marshmallow==3.0.0
marshmallow-sqlalchemy==0.17.0
python-dateutil==2.8.0
python-dotenv==0.10.3
python-editor==1.0.4
python-mpd2==1.0.0
six==1.12.0
SQLAlchemy==1.3.7
visitor==0.1.3
Werkzeug==0.15.5
WTForms==2.2.1
